+++ syncopy/datatype/__init__.py
@@ -13,6 +13,7 @@ from .methods.definetrial import *
 from .methods.padding import *
 from .methods.selectdata import *
 from .methods.show import *
+from .methods.copy import *
 
 # Populate local __all__ namespace
 __all__ = []
@@ -21,6 +22,8 @@ __all__.extend(continuous_data.__all__)
 __all__.extend(discrete_data.__all__)
 __all__.extend(statistical_data.__all__)
 __all__.extend(methods.definetrial.__all__)
-__all__.extend(methods.padding.__all__)
+# this is broken / has no current use case
+# __all__.extend(methods.padding.__all__)
 __all__.extend(methods.selectdata.__all__)
 __all__.extend(methods.show.__all__)
+__all__.extend(methods.copy.__all__)